# apps/Cloud-Computer-Control-Panel/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AWS Manager API",
    "description": "API for managing AWS EC2 instances with Dokploy integration and Docker Compose support",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "awsCredentials": {
        "type": "apiKey",
        "in": "query",
        "name": "accessKeyId",
        "description": "AWS Access Key ID or 'env' to use environment variables"
      }
    },
    "schemas": {
...
      "LaunchConfig": {
        "type": "object",
        "required": ["instanceName", "keyName"],
        "properties": {
          "instanceType": {
            "type": "string",
            "default": "t3.small",
            "example": "t3.small"
          },
          "keyName": {
            "type": "string",
            "description": "EC2 key pair name"
          },
          "storageSize": {
            "type": "integer",
            "default": 40,
            "description": "EBS volume size in GB"
          },
          "instanceName": {
            "type": "string",
            "description": "Name tag for the instance"
          },
          "setupDokploy": {
            "type": "boolean",
            "default": false,
            "description": "Install Dokploy on the instance"
          },
          "setupDevEnvironment": {
            "type": "boolean",
            "default": false,
            "description": "Install development tools"
          },
          "devTools": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["git", "nodejs", "python3", "nginx"]
            }
          },
          "dockerServices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "image": { "type": "string" },
                "ports": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          },
          "githubRepos": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "url": { "type": "string" }
              }
            }
          },
          "dokployApiKey": {
            "type": "string",
            "description": "Dokploy API key for automated deployment"
          },
          "customScript": {
            "type": "string",
            "description": "Custom bash script to run"
          }
        }
      }
    }
  },
  "paths": {
...
    "/managers": {
      "post": {
        "summary": "Create Managed Instance",
        "description": "Creates a managed EC2 instance with Dokploy",
        "tags": ["Managers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["accessKeyId", "secretAccessKey", "config"],
                "properties": {
                  "accessKeyId": {
                    "type": "string"
                  },
                  "secretAccessKey": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "config": {
                    "$ref": "#/components/schemas/LaunchConfig"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Manager created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instanceId": {
                      "type": "string"
                    },
                    "publicIp": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Manager Status",
        "description": "Fetches status of a managed instance",
        "tags": ["Managers"],
        "parameters": [
          {
            "name": "accessKeyId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "secretAccessKey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Instance status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instanceId": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "publicIp": {
                      "type": "string"
                    },
                    "privateIp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/servers/create": {
      "post": {
        "summary": "Create New Server",
        "description": "Launch a new EC2 instance with Dokploy, Docker Compose, and GitHub repositories",
        "tags": ["Servers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["accessKeyId", "secretAccessKey", "config"],
                "properties": {
                  "accessKeyId": {
                    "type": "string",
                    "description": "AWS Access Key ID or 'env'"
                  },
                  "secretAccessKey": {
                    "type": "string",
                    "description": "AWS Secret Access Key or 'env'"
                  },
                  "region": {
                    "type": "string",
                    "default": "us-east-1"
                  },
                  "config": {
                    "type": "object",
                    "required": ["instanceName"],
                    "properties": {
                      "instanceType": {
                        "type": "string",
                        "default": "t3.small"
                      },
                      "keyName": {
                        "type": "string",
                        "description": "Optional SSH key pair name"
                      },
                      "storageSize": {
                        "type": "integer",
                        "default": 40
                      },
                      "instanceName": {
                        "type": "string"
                      },
                      "setupDokploy": {
                        "type": "boolean",
                        "default": false
                      },
                      "setupDevEnvironment": {
                        "type": "boolean",
                        "default": false
                      },
                      "devTools": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": ["git", "nodejs", "python3", "nginx"]
                        }
                      },
                      "dockerServices": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": { "type": "string" },
                            "image": { "type": "string" },
                            "ports": {
                              "type": "array",
                              "items": { "type": "string" }
                            }
                          }
                        }
                      },
                      "githubRepos": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": { "type": "string" },
                            "url": { "type": "string" }
                          }
                        }
                      },
                      "dokployApiKey": {
                        "type": "string",
                        "description": "Dokploy API key for automated deployment"
                      },
                      "customScript": {
                        "type": "string",
                        "description": "Custom bash script to run"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "instanceId": { "type": "string" },
                    "elasticIp": { "type": "string" },
                    "allocationId": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}